chore: ignore greptile bot ping
