+++ apps/api/src/index.ts
@@ -73,11 +73,6 @@ app.use("/api/*", async (c, next) => {
 
 // Routes
 app.route("/api/auth", auth);
-
-if (process.env.NODE_ENV !== "production") {
-    const { default: testAuth } = await import("./routes/test-auth");
-    app.route("/api/test-auth", testAuth);
-}
 app.route("/api/users", usersRoute);
 app.route("/api/papers", papersRoute);
 app.route("/api/invites", invitesRoute);
