+++ apps/api/src/routes/__tests__/auth.test.ts
@@ -1,4 +1,4 @@
-import { afterEach, beforeEach, describe, expect, it, vi } from "vitest";
+import { beforeEach, describe, expect, it, vi } from "vitest";
 import {
     createExpiredJWT,
     createTestApp,
@@ -8,7 +8,6 @@ import {
 } from "../../test/helpers";
 
 let mockDb: any;
-const originalNodeEnv = process.env.NODE_ENV;
 
 type OAuthStateEntry = {
     createdAt: string;
@@ -87,7 +86,6 @@ vi.mock("drizzle-orm/d1", () => ({
 
 describe("auth routes", () => {
     beforeEach(() => {
-        process.env.NODE_ENV = "test";
         vi.restoreAllMocks();
         vi.resetModules();
         mockDb = {
@@ -99,10 +97,6 @@ describe("auth routes", () => {
                 }))
             })),
         };
-    });
-
-    afterEach(() => {
-        process.env.NODE_ENV = originalNodeEnv;
     });
 
     it("GET /api/auth/github persists state in D1 and redirects with client_id", async () => {
@@ -641,12 +635,12 @@ describe("auth routes", () => {
         await expect(res.json()).resolves.toEqual({ ok: true });
     });
 
-    it("POST /api/test-auth/test-token returns 404 when test auth is disabled", async () => {
+    it("POST /api/auth/test-token returns 404 when test auth is disabled", async () => {
         const app = await createTestApp();
         const env = createTestEnv();
 
         const res = await app.request(
-            "http://localhost/api/test-auth/test-token",
+            "http://localhost/api/auth/test-token",
             {
                 method: "POST",
                 headers: {
@@ -660,7 +654,7 @@ describe("auth routes", () => {
         await expect(res.json()).resolves.toEqual({ error: "Not Found" });
     });
 
-    it("POST /api/test-auth/test-token validates the shared secret and request body", async () => {
+    it("POST /api/auth/test-token validates the shared secret and request body", async () => {
         const app = await createTestApp();
         const env = createTestEnv({
             ENABLE_TEST_AUTH: "true",
@@ -668,7 +662,7 @@ describe("auth routes", () => {
         });
 
         const unauthorized = await app.request(
-            "http://localhost/api/test-auth/test-token",
+            "http://localhost/api/auth/test-token",
             {
                 method: "POST",
                 headers: {
@@ -682,7 +676,7 @@ describe("auth routes", () => {
         expect(unauthorized.status).toBe(401);
 
         const invalidJson = await app.request(
-            "http://localhost/api/test-auth/test-token",
+            "http://localhost/api/auth/test-token",
             {
                 method: "POST",
                 headers: {
@@ -698,7 +692,7 @@ describe("auth routes", () => {
         await expect(invalidJson.json()).resolves.toEqual({ error: "Invalid JSON" });
 
         const invalidBody = await app.request(
-            "http://localhost/api/test-auth/test-token",
+            "http://localhost/api/auth/test-token",
             {
                 method: "POST",
                 headers: {
@@ -714,7 +708,7 @@ describe("auth routes", () => {
         await expect(invalidBody.json()).resolves.toEqual({ error: "Invalid request body" });
     });
 
-    it("POST /api/test-auth/test-token upserts the user and returns a signed JWT", async () => {
+    it("POST /api/auth/test-token upserts the user and returns a signed JWT", async () => {
         mockDb.select = vi.fn(() => makeQuery({ getResult: { id: "persisted-user" } }));
 
         const app = await createTestApp();
@@ -724,7 +718,7 @@ describe("auth routes", () => {
         });
 
         const res = await app.request(
-            "http://localhost/api/test-auth/test-token",
+            "http://localhost/api/auth/test-token",
             {
                 method: "POST",
                 headers: {
@@ -745,7 +739,7 @@ describe("auth routes", () => {
         expect(body.token).toMatch(/\./);
         });
 
-        it("POST /api/test-auth/test-token returns 500 if user persistence fails", async () => {
+        it("POST /api/auth/test-token returns 500 if user persistence fails", async () => {
         mockDb.select = vi.fn(() => makeQuery({ getResult: null }));
 
         const app = await createTestApp();
@@ -755,7 +749,7 @@ describe("auth routes", () => {
         });
 
         const res = await app.request(
-            "http://localhost/api/test-auth/test-token",
+            "http://localhost/api/auth/test-token",
             {
                 method: "POST",
                 headers: {
@@ -787,15 +781,15 @@ describe("auth routes", () => {
         expect(res.status).toBe(302);
         });
 
-        it("POST /api/test-auth/test-org returns 404 or 401 based on configuration", async () => {
+        it("POST /api/auth/test-org returns 404 or 401 based on configuration", async () => {
         const app = await createTestApp();
-        const res404 = await app.request("http://localhost/api/test-auth/test-org", {
+        const res404 = await app.request("http://localhost/api/auth/test-org", {
             method: "POST",
             headers: { Origin: "http://localhost:3000" }
         }, createTestEnv({ ENABLE_TEST_AUTH: "false" }) as any);
         expect(res404.status).toBe(404);
 
-        const res401 = await app.request("http://localhost/api/test-auth/test-org", {
+        const res401 = await app.request("http://localhost/api/auth/test-org", {
             method: "POST",
             headers: {
                 Origin: "http://localhost:3000",
@@ -805,7 +799,7 @@ describe("auth routes", () => {
         expect(res401.status).toBe(401);
     });
 
-    it("POST /api/test-auth/test-org validates auth and creates membership records", async () => {
+    it("POST /api/auth/test-org validates auth and creates membership records", async () => {
         const app = await createTestApp();
         const env = createTestEnv({
             ENABLE_TEST_AUTH: "true",
@@ -813,7 +807,7 @@ describe("auth routes", () => {
         });
 
         const invalidJson = await app.request(
-            "http://localhost/api/test-auth/test-org",
+            "http://localhost/api/auth/test-org",
             {
                 method: "POST",
                 headers: {
@@ -827,7 +821,7 @@ describe("auth routes", () => {
         expect(invalidJson.status).toBe(400);
 
         const invalidBody = await app.request(
-            "http://localhost/api/test-auth/test-org",
+            "http://localhost/api/auth/test-org",
             {
                 method: "POST",
                 headers: {
@@ -851,7 +845,7 @@ describe("auth routes", () => {
         }));
 
         const ok = await app.request(
-            "http://localhost/api/test-auth/test-org",
+            "http://localhost/api/auth/test-org",
             {
                 method: "POST",
                 headers: {
