+++ apps/api/src/routes/collections.ts
@@ -16,7 +16,6 @@ import {
 import { authMiddleware } from "../middleware/auth";
 import type { Env, Variables } from "../types";
 import { validateSlug, validateName, validateDescription } from "../utils/validation";
-import { getOrgBySlug, getOrgMembership, isOrgMember, isOrgAdmin, isPaperAuthor, isMemberOfPaperOrg } from "../utils/db";
 
 const collectionsRoute = new Hono<{ Bindings: Env; Variables: Variables }>();
 
@@ -52,6 +51,27 @@ async function getCurrentUser(c: any): Promise<CurrentUser> {
     } catch {
         return null;
     }
+}
+
+async function getOrgBySlug(db: ReturnType<typeof drizzle>, slug: string) {
+    return db.select().from(orgs).where(eq(orgs.slug, slug)).get();
+}
+
+async function getOrgMembership(db: ReturnType<typeof drizzle>, orgId: string, userId: string) {
+    return db
+        .select()
+        .from(orgMembers)
+        .where(and(eq(orgMembers.orgId, orgId), eq(orgMembers.userId, userId)))
+        .get();
+}
+
+async function isOrgMember(db: ReturnType<typeof drizzle>, orgId: string, userId: string) {
+    return !!(await getOrgMembership(db, orgId, userId));
+}
+
+async function isOrgAdmin(db: ReturnType<typeof drizzle>, orgId: string, userId: string) {
+    const row = await getOrgMembership(db, orgId, userId);
+    return !!row && (row.role === "admin" || row.role === "owner");
 }
 
 async function canViewCollection(
@@ -92,10 +112,23 @@ async function canViewPaper(
     if (paper.visibility === "public") return true;
     if (!userId) return false;
 
-    if (await isPaperAuthor(db, paper.id, userId)) return true;
+    const isAuthor = await db
+        .select({ paperId: paperAuthors.paperId })
+        .from(paperAuthors)
+        .where(and(eq(paperAuthors.paperId, paper.id), eq(paperAuthors.userId, userId)))
+        .get();
+    if (isAuthor) return true;
+
     if (paper.visibility === "private") return false;
 
-    return await isMemberOfPaperOrg(db, paper.id, userId);
+    // org_only: user must be a member of an org that owns this paper
+    const membership = await db
+        .select({ orgId: orgMembers.orgId })
+        .from(orgMembers)
+        .innerJoin(paperOrgs, eq(orgMembers.orgId, paperOrgs.orgId))
+        .where(and(eq(paperOrgs.paperId, paper.id), eq(orgMembers.userId, userId)))
+        .get();
+    return !!membership;
 }
 
 collectionsRoute.post("/collections", authMiddleware, async (c) => {
