+++ apps/e2e/tests/papers.spec.ts
@@ -277,7 +277,7 @@ test.describe('論文ダウンロード', () => {
         const authSecret = process.env.TEST_AUTH_SECRET as string;
 
         const apiURL = process.env.E2E_API_URL || 'http://localhost:8787';
-        const setupRes = await page.request.post(`${apiURL}/api/test-auth/test-org`, {
+        const setupRes = await page.request.post(`${apiURL}/api/auth/test-org`, {
             headers: { 'x-test-auth-secret': authSecret },
             data: { userId: memberUserId, orgId }
         });
